(PC-11354) refactor(AppSearch): remove code related to offers
