+++ src/features/home/pages/tests/Home.native.test.tsx
@@ -7,9 +7,9 @@ import { UserProfileResponse } from 'api/gen'
 import { AuthContext } from 'features/auth/AuthContext'
 import { ProcessedModule } from 'features/home/contentful'
 import { RecommendationPane } from 'features/home/contentful/moduleTypes'
+import { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
 import { analytics } from 'libs/analytics'
 import { env } from 'libs/environment'
-import { mockedAlgoliaResponse } from 'libs/search/fixtures'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
 import { flushAllPromises, superFlushWithAct, render, act } from 'tests/utils'
