--- src/libs/search/filters/index.ts
@@ -1,2 +1,0 @@
-export { AppSearchFields, RESULT_FIELDS } from './constants'
-export { underageFilter, buildQueryOptions } from './buildQueryOptions'
