--- src/libs/search/fixtures/index.tsx
@@ -1,1 +1,0 @@
-export { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
